+++ nbrmd/__init__.py
@@ -9,4 +9,4 @@ Use the 'nbrmd' conversion script to convert Jupyter notebooks from/to R markdow
 """
 
 from .nbrmd import read, reads, readf, write, writes, writef
-from .jupyter_hook import pre_save_hook
+from .hooks import update_rmd, update_ipynb, update_rmd_and_ipynb